major refactor for simplicity
