+++ .env
@@ -1,6 +1,3 @@
 DB_PASSWORD = 
 PMA_PASSWORD = 
-
-STUDENT_SCHOOL_DB = 
-BOOK_DB = 
-
+SQLALCHEMY_DATABASE_URI =
