+++ backend/DOCKERFILE
@@ -1,0 +1,9 @@
+FROM python:3-slim
+
+WORKDIR /backend
+
+COPY requirements.txt .
+
+RUN pip install --no-cache-dir -r requirements.txt
+
+COPY . .
