--- backend/services/book/book.DOCKERFILE
@@ -1,12 +1,0 @@
-FROM python:3.9-slim
-
-WORKDIR /backend/services/book
-
-COPY requirements.txt .
-
-RUN apt-get update \ 
-    && pip install -r requirements.txt
-
-COPY . .
-
-CMD ["python", "./book_api.py"]
