+++ docker-compose.yml
@@ -1,33 +1,20 @@
 version: "3.9"
 
-services:
-  ###################################
-  # Student_School: The Student_School microservice
-  ###################################
-  student_school:
+services:  
+  webserver:
     build:
-      context: ./backend/services/student_school
-      dockerfile: student_school.DOCKERFILE
-    image: student_school
-    restart: always
-    environment:
-      SQLALCHEMY_DATABASE_URI: ${STUDENT_SCHOOL_DB}
-    ports:
-      - 6000:6000
-  
-  ###################################
-  # Book: The book microservice
-  ###################################
-  book:
-    build:
-      context: ./backend/services/book
-      dockerfile: book.DOCKERFILE
-    image: book
-    restart: always
-    environment:
-      SQLALCHEMY_DATABASE_URI: ${BOOK_DB}
+      context: ./backend
+      dockerfile: DOCKERFILE
+    command: python app.py
     ports:
       - 6010:6010
+    depends_on:
+      - db
+    restart: always
+    volumes:
+      - "./backend:/backend"
+    environment:
+      SQLALCHEMY_DATABASE_URI: ${SQLALCHEMY_DATABASE_URI}
   
   db:
     image: mysql
@@ -36,6 +23,7 @@ services:
     environment:
       MYSQL_ROOT_PASSWORD: root
       MYSQL_PASSWORD: ${DB_PASSWORD}
+      MYSQL_DATABASE: student_life_db
     volumes:             
       - ./backend/init:/docker-entrypoint-initdb.d
     ports:
